1.修复邀请注册bug

+++ user/service/src/main/java/com/yryz/quanhu/user/service/UserOperateService.java
@@ -58,7 +58,7 @@ public interface UserOperateService {
 	 * @return
 	 */
 	MyInviterVO getMyInviter(Long userId, Integer limit,
-			Integer inviterId);
+			Long inviterId);
 
 	/**
 	 * 查询注册信息
